Cleanup 2_affine_transforms notebook (#195)
Cleanup 2_affine_transforms.md to not have any duplicate code with
python modules and present it via `{literalinclude}` in the
documentation.
Restructure plotting and point modules to break them down into
submodules that can be imported independently for less footprint and
make it easier to display them in the docs
