+++ docs/source/notebooks/geometry/point.py
@@ -76,23 +76,6 @@ class Point:
         return self / np.linalg.norm(self.numpy())
 
 
-class Line:
-    """A simple 2D line class."""
-
-    def __init__(self, start: Point, end: Point, label: str):
-        self.start = start
-        self.end = end
-        self.label = label
-
-    def extended(self, length: float = 1.0):
-        return Line(
-            self.start, self.end + (self.end - self.start).normalized() * length, self.label
-        )
-
-    def __repr__(self):
-        return f'Line({self.start}, {self.end}, {self.label})'
-
-
 class SimplePoint3D:
     """A simple 3D point class for getting_started_with_robot_ik."""
 
@@ -232,51 +215,3 @@ class Point3D:
 
     def copy(self):
         return Point3D(self._array, self.label)
-
-
-class Line3D:
-    """A simple 3D line class."""
-
-    def __init__(self, start: Point3D, end: Point3D, label: str):
-        self.start = start
-        self.end = end
-        self.label = label
-
-    def extended(self, length: float = 1.0):
-        return Line3D(
-            self.start, self.end + (self.end - self.start).normalized() * length, self.label
-        )
-
-    @property
-    def xy(self):
-        return Line(self.start.xy, self.end.xy, self.label)
-
-    @property
-    def xz(self):
-        return Line(self.start.xz, self.end.xz, self.label)
-
-    @property
-    def yz(self):
-        return Line(self.start.yz, self.end.yz, self.label)
-
-
-class Leg3D:
-    """A 3D leg class."""
-
-    def __init__(self, lines: list[Line3D]):
-        self.lines = lines
-
-    def __iter__(self):
-        return iter(self.lines)
-
-    @property
-    def xy(self):
-        return [line.xy for line in self.lines]
-
-    @property
-    def xz(self):
-        return [line.xz for line in self.lines]
-
-    @property
-    def yz(self):
-        return [line.yz for line in self.lines]
