+++ docs/source/notebooks/jupyter_utils.py
@@ -4,15 +4,26 @@ from pygments.formatters import HtmlFormatter
 from pygments.lexers import PythonLexer
 
 
-def display_file(file_path, start_line=0, end_line=-1, style='default'):
+def display_file(
+    file_path, start_line=-1, end_line=-1, start_after=None, end_before=None, style='default'
+):
     with open(file_path) as f:
-        code = f.readlines()[start_line:end_line]
+        code = f.readlines()
+        if start_line > 0:
+            code = code[start_line:end_line]
         code_str = ''.join(code)
+
+        if start_after:
+            end_index = -1
+            if end_before:
+                end_index = code_str.index(end_before)
+
+            code_str = code_str[code_str.index(start_after) + len(start_after) : end_index]
 
     formatter = HtmlFormatter(style=style)
     return IPython.display.HTML(
         f'<style type="text/css">{formatter.get_style_defs()}</style>'
-        + f'<h3>"{file_path}", lines {start_line} to {end_line}</h3>'
+        # + f'<h3>"{file_path}", lines {start_line} to {end_line}</h3>'
         + highlight(code_str, PythonLexer(), formatter)
     )
