+++ docs/source/notebooks/requirements.txt
@@ -10,4 +10,4 @@ typeguard>=4.1.5  # Required by generate-parameter-library-py
 
 Pygments>=2.17.2
 
-jupytext>=1.17.0
+jupytext==1.17.2
